Backend/Grad_Project/requirements.txt
Flask==2.2.1
Flask-RESTful==0.3.9
gunicorn==20.1.0
bcrypt==3.2.2
blueprint==3.4.2
pandas==1.4.3
PyPDF2==2.10.5
pydub==0.25.1
tokenlib==2.0.0
google-cloud==0.34.0
firebase==3.0.1
firebase-admin==5.2.0
google-cloud-vision==3.1.1
google-cloud-storage==2.4.0
oauthlib==3.2.1
python-jwt==3.3.3
python-dotenv==0.21.0
path==16.4.0
pandas==1.4.3
PyJWT==2.4.0
uuid==1.30
Flask-Cors==3.0.10
google-cloud-translate==3.8.0